+++ app/src/main/cpp/CMakeLists.txt
@@ -37,6 +37,7 @@ add_library( # Sets the name of the library.
         shader.cpp
         framebuffer.cpp
         data_texture_pair.cpp
+        wavelet_turbulence.cpp
         )
 
 # Searches for a specified prebuilt library and stores the path as a
