+++ app/src/main/cpp/helper.h
@@ -14,16 +14,6 @@ using namespace glm;
 
 bool checkGlError(const char *funcName);
 
-double fade(double x);
-
-double lerp(double a, double b, double t);
-
-double perlin(double x, double y, double z);
-
-vec4 noise(double x, double y, double z);
-
-void generate3DTexture(GLuint *textureID, GLsizei width, GLsizei height, GLsizei depth);
-
 void createScalar3DTexture(GLuint *id, int width, int height, int depth, float* data);
 void createVector3DTexture(GLuint *id, int width, int height, int depth, vec3* data);
