+++ app/src/main/cpp/wavelet_turbulence.h
@@ -1,0 +1,31 @@
+//
+// Created by Kalle on 2020-04-01.
+//
+
+#ifndef DATX02_20_21_WAVELET_TURBULENCE_H
+#define DATX02_20_21_WAVELET_TURBULENCE_H
+
+#include <gles3/gl31.h>
+#include <GLES3/gl3ext.h>
+
+#include <glm/glm.hpp>
+#include <glm/gtc/matrix_transform.hpp>
+
+#include <android/log.h>
+
+#define LOG_TAG "Renderer"
+#define ALOGE(...) __android_log_print(ANDROID_LOG_ERROR, LOG_TAG, __VA_ARGS__)
+
+using namespace glm;
+
+#define PI 3.14159265359f
+
+class WaveletTurbulence {
+public:
+    double turbulance(vec3 position);
+    double perlin(vec3(position));
+private:
+};
+
+
+#endif //DATX02_20_21_WAVELET_TURBULENCE_H
